RPM packages build again.

+++ hcore/hvariant.hxx
@@ -83,7 +83,7 @@ private:
 	typedef typename yaal::trait::ternary<( SIZE > ALIGNOF_INT_LONG ), int long long,
 					typename yaal::trait::ternary<( SIZE > ALIGNOF_INT ), int long,
 					typename yaal::trait::ternary<( SIZE > ALIGNOF_INT_SHORT ), int,
-					typename yaal::trait::ternary<( SIZE > sizeof ( char ) ), int short, char>::type>::type>::type>::type aligner_t;
+					typename yaal::trait::ternary<( SIZE > static_cast<int>( sizeof ( char ) ) ), int short, char>::type>::type>::type>::type aligner_t;
 	static int const ALIGNER_COUNT = SIZE / static_cast<int>( sizeof ( aligner_t ) ) + ( SIZE % static_cast<int>( sizeof ( aligner_t ) ) ? 1 : 0 );
 	aligner_t _mem[ ALIGNER_COUNT ];
 	int _type;
